Tomato Freshness/Tomato Freshness/requirements.txt
flask>=2.3.0
tensorflow>=2.10.0
opencv-python>=4.7.0
numpy>=1.23.0
pillow>=9.3.0
